+++ NEWS.md
@@ -1,3 +1,26 @@
+# 0.0.13
+
+## New functions
+* NMexpandDoses - Transform repeated dosing events (ADDL/II) to
+  individual dosing events
+* addTAPD - Add cumulative number of doses, time of last dose,
+  previous dose amount, cumulative dose amount, and time since
+  previous dose to data
+* tmpcol provides column names not already used in data sets. tmpcol
+  has long been part of NMdata but has not been exported until now.
+
+## New data
+* A new data set called mad is included. It is based on the
+  mad_missing_duplicates from the xgxr package. Doses are implemented
+  using ADDL and II (so only one dosing row per subject). It is
+  included for testing the new NMexpandDoses and coming addTAD
+  functions.
+
+## Bugfixes
+* Non-critical bugfix in mergeCheck dimensions overview printed to
+  console. One column too many was reported in input and result
+  data. No implications on results from mergeCheck.
+
 # NMdata 0.0.12
 
 * Vignettes are no longer included in R package and can only be read